start packaging the code to be more package like
